Actualización de player, (nuevo) amigos y lógica del juego
Actualización del jugador - nuevos campos avatar y money
Delete - comprobar que no esté en un juego, excepción
Amigos - agregar, eliminar, leer perfil de amigo (player/read), consultar listaAmigos
Implementación de la lógica del juego - start game, playCard

+++ src/main/java/com/unizar/unozar/core/service/PlayerServiceImpl.java
@@ -1,5 +1,6 @@
 package com.unizar.unozar.core.service;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
@@ -9,6 +10,8 @@ import com.unizar.unozar.core.controller.resources.AuthenticationRequest;
 import com.unizar.unozar.core.controller.resources.AuthenticationResponse;
 import com.unizar.unozar.core.controller.resources.CreatePlayerRequest;
 import com.unizar.unozar.core.controller.resources.DeletePlayerRequest;
+import com.unizar.unozar.core.controller.resources.FriendListResponse;
+import com.unizar.unozar.core.controller.resources.FriendRequest;
 import com.unizar.unozar.core.controller.resources.ReadPlayerRequest;
 import com.unizar.unozar.core.controller.resources.TokenRequest;
 import com.unizar.unozar.core.controller.resources.TokenResponse;
@@ -17,6 +20,7 @@ import com.unizar.unozar.core.entities.Player;
 import com.unizar.unozar.core.exceptions.EmailInUse;
 import com.unizar.unozar.core.exceptions.InvalidIdentity;
 import com.unizar.unozar.core.exceptions.InvalidPassword;
+import com.unizar.unozar.core.exceptions.PlayerIsPlaying;
 import com.unizar.unozar.core.exceptions.PlayerNotFound;
 import com.unizar.unozar.core.repository.PlayerRepository;
 
@@ -74,6 +78,9 @@ public class PlayerServiceImpl implements PlayerService{
   public Void delete(DeletePlayerRequest request){
     Player toDelete = findPlayer(request.getToken().substring(0, 32));
     checkToken(toDelete, request.getToken().substring(32));
+    if (!toDelete.getGameId().equals(Player.NONE)){
+      throw new PlayerIsPlaying("The player is on a game");
+    }
     playerRepository.delete(toDelete);
     return null;
   }
@@ -104,7 +111,7 @@ public class PlayerServiceImpl implements PlayerService{
   
   public Player findPlayer(String id){ 
     Optional<Player> toFind = playerRepository.findById(id);
-    if (!toFind.isPresent()){
+    if(!toFind.isPresent()){
       throw new PlayerNotFound("Id does not exist in the system");
     }
     return toFind.get();
@@ -116,4 +123,56 @@ public class PlayerServiceImpl implements PlayerService{
     }
     return null;
   }
+
+  @Override
+  public TokenResponse addFriend(FriendRequest request){
+    Player toMakeFriend = findPlayer(request.getToken().substring(0, 32));
+    checkToken(toMakeFriend, request.getToken().substring(32));
+    Optional<Player> toAdd = playerRepository.findById(request.getFriendId());
+    if(!toAdd.isPresent()){
+      throw new PlayerNotFound("Id does not exist in the system");
+    }
+    toMakeFriend.addNewFriend(request.getFriendId());
+    String token = toMakeFriend.getId() + toMakeFriend.updateSession();
+    playerRepository.save(toMakeFriend);
+    return new TokenResponse(token);
+  }
+
+  @Override
+  public FriendListResponse readFriendList(TokenRequest request){
+    Player toReadFriends = findPlayer(request.getToken().substring(0, 32));
+    checkToken(toReadFriends, request.getToken().substring(32));
+    List<String> l = toReadFriends.getFriendList();
+    String[] alias = new String[l.size()];
+    String[] emails = new String[l.size()];
+    int[] avatarIds = new int[l.size()];
+    for(int i = 0; i < l.size(); i++){
+      Optional<Player> toFind = playerRepository.findById(l.get(i));
+      if(toFind.isPresent()){
+        Player toRead = toFind.get();
+        alias[i] = toRead.getAlias();
+        emails[i] = toRead.getEmail();
+        avatarIds[i] = toRead.getAvatarId();
+      }else{
+        toReadFriends.deleteFriend(l.get(i));
+        l.remove(i);
+        i--;
+      }
+    }
+    String token = toReadFriends.getId() + toReadFriends.updateSession();
+    playerRepository.save(toReadFriends);
+    return new FriendListResponse(token, l.toArray(new String[0]), alias, 
+        emails, avatarIds);
+  }
+  
+  @Override
+  public TokenResponse deleteFriend(FriendRequest request){
+    Player toDeleteFriend = findPlayer(request.getToken().substring(0, 32));
+    checkToken(toDeleteFriend, request.getToken().substring(32));
+    toDeleteFriend.deleteFriend(request.getFriendId());
+    String token = toDeleteFriend.getId() + toDeleteFriend.updateSession();
+    playerRepository.save(toDeleteFriend);
+    return new TokenResponse(token);
+  }
+
 }
